TransportHome Package Interface
* AppHome만 빌드하려 했을때 import하고 있는 TransportHome, Finance와 같은 다른 패키지를 빌드해야한다
* 이때, TransportHome이 가진 Builder, Interactor, ViewController .. 등등 모두 빌드하게됨 (시간이 오래걸린다)
* AppHome 입장에서는 TransportHome을 다 빌드할 필요가 없기때문에, TransportHome의 Interface만 빌드하게 하고싶다
* TransportHome -> Interface만 제공하는 패키지 / TransportHomeImp -> 실제 비즈니스의 구현 패키지

* 모듈간의 의존성이 엮이는 부분은 Builder임 -> Builder에서의 느슨한 결합을 통해 각 모듈들을 분리시키고자함
	* AppHomeComponent가 TransportHomeDependency를 채택하지 않고 부모 Dependency로부터 주입받도록 함
	* 기존 build 메소드에서 TransportHomeBuilder(imp)를 생성하지 않게한다
	* AppRootComponent가 TransportHomeDependency를 가지고 TransportHomeBuilder(imp)를 만들어줌
	* 이로써 AppHome입장에서는 위에서 내려오는 transportHomeBuildable을 사용할거기때문에 의존성이 사라짐
	* AppHome이 빌드 될때 TransportHome의 Interface만 필요로 하게됨 -> 시간이 감소한다, AppHome만 확인할 수 있게 된다
