+++ SuperApp/AppRoot/AppRootBuilder.swift
@@ -17,22 +17,6 @@ protocol AppRootDependency: Dependency {
     // created by this RIB.
 }
 
-final class AppRootComponent: Component<AppRootDependency>, AppHomeDependency, FinanceHomeDependency, ProfileHomeDependency {
-    var cardOnFileRepository: CardOnFileRepository
-    
-    var superPayRepository: SuperPayRepository
-    
-    init(
-        dependency: AppRootDependency,
-        cardOnFileRepository: CardOnFileRepository,
-        superPayRepository: SuperPayRepository
-    ) {
-        self.cardOnFileRepository = cardOnFileRepository
-        self.superPayRepository = superPayRepository
-        super.init(dependency: dependency)
-    }
-}
-
 // MARK: - Builder
 
 protocol AppRootBuildable: Buildable {
