+++ Transport/Package.swift
@@ -11,6 +11,10 @@ let package = Package(
             name: "TransportHome",
             targets: ["TransportHome"]
         ),
+        .library(
+            name: "TransportHomeImp",
+            targets: ["TransportHomeImp"]
+        ),
     ],
     dependencies: [
         .package(name: "ModernRIBs", url: "https://github.com/DevYeom/ModernRIBs", .exact("1.0.1")),
@@ -21,7 +25,14 @@ let package = Package(
         .target(
             name: "TransportHome",
             dependencies: [
+                "ModernRIBs"
+            ]
+        ),
+        .target(
+            name: "TransportHomeImp",
+            dependencies: [
                 "ModernRIBs",
+                "TransportHome",
                 .product(name: "FinanceRepository", package: "Finance"),
                 .product(name: "Topup", package: "Finance"),
             ],
