+++ Transport/Sources/TransportHomeImp/TransportHomeBuilder.swift
@@ -4,6 +4,7 @@ import FinanceRepository
 import CombineUtil
 import FoundationUtil
 import Topup
+import TransportHome
 
 public protocol TransportHomeDependency: Dependency {
     var cardOnFileRepository: CardOnFileRepository { get }
@@ -32,10 +33,6 @@ final class TransportHomeComponent: Component<TransportHomeDependency>, Transpor
 }
 
 // MARK: - Builder
-
-public protocol TransportHomeBuildable: Buildable {
-    func build(withListener listener: TransportHomeListener) -> ViewableRouting
-}
 
 public final class TransportHomeBuilder: Builder<TransportHomeDependency>, TransportHomeBuildable {
