+++ Transport/Sources/TransportHomeImp/TransportHomeInteractor.swift
@@ -2,6 +2,7 @@ import ModernRIBs
 import Combine
 import Foundation
 import CombineUtil
+import TransportHome
 
 protocol TransportHomeRouting: ViewableRouting {
     func attachTopup()
@@ -12,10 +13,6 @@ protocol TransportHomePresentable: Presentable {
     var listener: TransportHomePresentableListener? { get set }
     
     func setSuperPayBalance(_ balanceText: String)
-}
-
-public protocol TransportHomeListener: AnyObject {
-    func transportHomeDidTapClose()
 }
 
 protocol TransportHomeInteractorDependency {
